+++ oxe-api/test/resource/user/test_update_user.py
@@ -15,7 +15,8 @@ class TestUpdateUser(BaseCase):
 
         payload = {
             "id": 14,
-            "is_admin": True
+            "is_admin": True,
+            "email": "mynewemail@test.lu"
         }
 
         response = self.application.post('/user/update_user',
@@ -27,6 +28,7 @@ class TestUpdateUser(BaseCase):
         self.assertEqual(200, response.status_code)
         self.assertEqual(len(users), 1)
         self.assertEqual(users[0].is_admin, 1)
+        self.assertEqual(users[0].email, "mynewemail@test.lu")
 
     @BaseCase.login
     @BaseCase.grant_access("/user/update_user")
@@ -55,7 +57,7 @@ class TestUpdateUser(BaseCase):
 
     @BaseCase.login
     @BaseCase.grant_access("/user/update_user")
-    def test_ko_email_param(self, token):
+    def test_ko_with_wrong_email(self, token):
         self.db.insert({
             "id": 2,
             "email": "myemail@test.lu",
@@ -66,7 +68,7 @@ class TestUpdateUser(BaseCase):
         payload = {
             "id": 2,
             "is_admin": True,
-            "email": "myemail@test.lu"
+            "email": "myemaildeded@dde@dedde"
         }
 
         response = self.application.post('/user/update_user',
@@ -77,3 +79,4 @@ class TestUpdateUser(BaseCase):
 
         self.assertEqual("422 UNPROCESSABLE ENTITY", response.status)
         self.assertEqual(users[0].is_admin, 0)
+        self.assertEqual(users[0].email, "myemail@test.lu")